requirements.txt
# Core UI Framework
PyQt6>=6.6.0

# Async support for Qt
qasync>=0.27.0

# PDF handling
PyMuPDF>=1.23.0
pdf2image
pillow

# EPUB handling
ebooklib>=0.18

# Dictionary support
nltk>=3.8.1

# Encryption
cryptography>=42.0.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
pylint>=3.0.0
mypy>=1.7.0